Enclave/requirements.txt
pycryptodome==3.17
pyOpenSSL==23.1.1
requests==2.24.0
